refactor: Address PR comments
