+++ src/components/Wallet/WalletBridges.tsx
@@ -1,9 +1,9 @@
 import useChainConfig from 'hooks/chain/useChainConfig'
-import CosmosWalletBridges from './CosmosWalletBridges'
-import NeutronEVMWalletBridges from './NeutronEVMWalletBridges'
+import CosmosWalletBridges from 'components/Wallet/CosmosWalletBridges'
+import EvmSupportedBridges from 'components/Wallet/EvmSupportedBridges'
 
 export default function WalletBridges() {
   const chainConfig = useChainConfig()
 
-  return chainConfig.evmAssetSupport ? <NeutronEVMWalletBridges /> : <CosmosWalletBridges />
+  return chainConfig.evmAssetSupport ? <EvmSupportedBridges /> : <CosmosWalletBridges />
 }
