+++ src/constants/ethereumConfig.ts
